Refactor QCJob.py
Refactored QCJob.py, integrating the code from QCHelper and QCCmdGenerator into a single class.
